+++ packages/alchemy/src/index.ts
@@ -12,6 +12,8 @@ export type * from "./client/lightAccountClient.js";
 export { createLightAccountAlchemyClient } from "./client/lightAccountClient.js";
 export type * from "./client/modularAccountClient.js";
 export { createModularAccountAlchemyClient } from "./client/modularAccountClient.js";
+export type * from "./client/multisigAccountClient.js";
+export { createMultisigAccountAlchemyClient } from "./client/multisigAccountClient.js";
 export type * from "./client/rpcClient.js";
 export { createAlchemyPublicRpcClient } from "./client/rpcClient.js";
 export type * from "./client/smartAccountClient.js";
@@ -25,6 +27,11 @@ export { alchemyUserOperationSimulator } from "./middleware/userOperationSimulat
 export type * from "./schema.js";
 export { AlchemyProviderConfigSchema } from "./schema.js";
 export type { AlchemyProviderConfig } from "./type.js";
+export { getAlchemyPaymasterAddress } from "./gas-manager.js";
 
 export type * from "./signer/index.js";
-export { AlchemySigner, AlchemySignerClient } from "./signer/index.js";
+export {
+  AlchemySigner,
+  AlchemySignerClient,
+  AlchemySignerStatus,
+} from "./signer/index.js";