+++ packages/alchemy/src/middleware/gasManager.ts
@@ -1,4 +1,5 @@
 import type {
+  Address,
   ClientMiddlewareConfig,
   ClientMiddlewareFn,
 } from "@alchemy/aa-core";
@@ -16,6 +17,7 @@ import {
 } from "@alchemy/aa-core";
 import { fromHex } from "viem";
 import type { ClientWithAlchemyMethods } from "../client/types";
+import { getAlchemyPaymasterAddress } from "../gas-manager.js";
 import { alchemyFeeEstimator } from "./feeEstimator.js";
 
 export type RequestGasAndPaymasterAndDataOverrides = Partial<{
@@ -29,6 +31,8 @@ export type RequestGasAndPaymasterAndDataOverrides = Partial<{
 export interface AlchemyGasManagerConfig {
   policyId: string;
   gasEstimationOptions?: AlchemyGasEstimationOptions;
+  paymasterAddress?: Address;
+  dummyData?: Hex;
 }
 
 export interface AlchemyGasEstimationOptions {
@@ -38,17 +42,18 @@ export interface AlchemyGasEstimationOptions {
 }
 
 const dummyPaymasterAndData =
-  <C extends ClientWithAlchemyMethods>(client: C) =>
+  <C extends ClientWithAlchemyMethods>(
+    client: C,
+    config: AlchemyGasManagerConfig
+  ) =>
   () => {
-    switch (client.chain.id) {
-      case 1:
-      case 10:
-      case 137:
-      case 42161:
-        return "0x4Fd9098af9ddcB41DA48A1d78F91F1398965addcfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000007aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa1c";
-      default:
-        return "0xc03aac639bb21233e0139381970328db8bceeb67fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000007aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa1c";
-    }
+    const paymasterAddress =
+      config.paymasterAddress ?? getAlchemyPaymasterAddress(client.chain);
+    const dummyData =
+      config.dummyData ??
+      "fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000007aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa1c";
+
+    return `${paymasterAddress}${dummyData}` as Address;
   };
 
 export const alchemyGasManagerMiddleware = <C extends ClientWithAlchemyMethods>(
@@ -84,6 +89,7 @@ export const alchemyGasManagerMiddleware = <C extends ClientWithAlchemyMethods>(
                 overrides,
                 account,
                 feeOptions,
+                client,
               }),
             };
           }
@@ -108,6 +114,7 @@ export const alchemyGasManagerMiddleware = <C extends ClientWithAlchemyMethods>(
               overrides,
               feeOptions,
               account,
+              client,
             });
             maxFeePerGas = (await result.maxFeePerGas) ?? maxFeePerGas;
             maxPriorityFeePerGas =
@@ -130,7 +137,7 @@ const requestGasAndPaymasterData: <C extends ClientWithAlchemyMethods>(
   client: C,
   config: AlchemyGasManagerConfig
 ) => ClientMiddlewareConfig["paymasterAndData"] = (client, config) => ({
-  dummyPaymasterAndData: dummyPaymasterAndData(client),
+  dummyPaymasterAndData: dummyPaymasterAndData(client, config),
   paymasterAndData: async (struct, { overrides, feeOptions, account }) => {
     const userOperation: UserOperationRequest = deepHexlify(
       await resolveProperties(struct)
@@ -199,7 +206,7 @@ const requestPaymasterAndData: <C extends ClientWithAlchemyMethods>(
   client: C,
   config: AlchemyGasManagerConfig
 ) => ClientMiddlewareConfig["paymasterAndData"] = (client, config) => ({
-  dummyPaymasterAndData: dummyPaymasterAndData(client),
+  dummyPaymasterAndData: dummyPaymasterAndData(client, config),
   paymasterAndData: async (struct, { account }) => {
     const { paymasterAndData } = await client.request({
       method: "alchemy_requestPaymasterAndData",